Rename FieldList to FieldsOverview
