+++ graylog2-web-interface/src/views/components/sidebar/index.jsx
@@ -1,5 +1,5 @@
 export { default as AddWidgetButton } from './AddWidgetButton';
-export { default as FieldList } from './fields/FieldList';
+export { default as FieldsOverview } from './fields/FieldsOverview';
 export { default as SearchResultOverview } from './SearchResultOverview';
 // eslint-disable-next-line import/no-cycle
 export { default as SideBar } from './SideBar';
